project/learning_openai_connect/pyproject.toml
[project]
name = "learning-openai-connect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.85.0",
    "python-dotenv>=1.1.0",
]
